Refactor Scheduler to use a LinkedHashSet
Instead of maintaining 3 fields:

Hashtable m_commandTable
LinkedListElement m_firstCommand
LinkedListElement m_lastCommand

and maintaining a Linked List and a Hash table of all of the running
commands, use the library data structure LinkedHashSet, which already
does all of this.  

LinkedHashSet guarantees that elements will be returned in the order
of addition.